Moved all methods in HappyHexGUI.java to GameEssentials.java
