+++ Launcher/LaunchEssentials.java
@@ -47,23 +47,23 @@ public final class LaunchEssentials {
     public static void startGame(){
         gameStarted = true;
         if (currentGameInfo.getGameMode() == GameMode.Small){
-            HappyHexGUI.initialize(5, 3, 100, false, LauncherGUI.mainFrame);
+            GameEssentials.initialize(5, 3, 100, false, LauncherGUI.mainFrame);
         } else if (currentGameInfo.getGameMode() == GameMode.Medium){
-            HappyHexGUI.initialize(8, 5, 100, false, LauncherGUI.mainFrame);
+            GameEssentials.initialize(8, 5, 100, false, LauncherGUI.mainFrame);
         } else if (currentGameInfo.getGameMode() == GameMode.Large){
-            HappyHexGUI.initialize(11, 7, 100, false, LauncherGUI.mainFrame);
+            GameEssentials.initialize(11, 7, 100, false, LauncherGUI.mainFrame);
         } else if (currentGameInfo.getGameMode() == GameMode.SmallEasy){
-            HappyHexGUI.initialize(5, 3, 100, true, LauncherGUI.mainFrame);
+            GameEssentials.initialize(5, 3, 100, true, LauncherGUI.mainFrame);
         } else if (currentGameInfo.getGameMode() == GameMode.MediumEasy){
-            HappyHexGUI.initialize(8, 5, 100, true, LauncherGUI.mainFrame);
+            GameEssentials.initialize(8, 5, 100, true, LauncherGUI.mainFrame);
         } else if (currentGameInfo.getGameMode() == GameMode.LargeEasy){
-            HappyHexGUI.initialize(11, 7, 100, true, LauncherGUI.mainFrame);
+            GameEssentials.initialize(11, 7, 100, true, LauncherGUI.mainFrame);
         } else if(currentGameInfo.getGameMode() == GameMode.Unspecified){
             System.err.println("Legacy GameMode.Unspecified GameMode unsupported since Version 0.4.1");
-            HappyHexGUI.initialize(5, 3, 100, false, LauncherGUI.mainFrame);
+            GameEssentials.initialize(5, 3, 100, false, LauncherGUI.mainFrame);
         } else {
             System.err.println("Unknown GameMode detected.");
-            HappyHexGUI.initialize(5, 3, 100, false, LauncherGUI.mainFrame);
+            GameEssentials.initialize(5, 3, 100, false, LauncherGUI.mainFrame);
         }
     }
     public static void endGame(){
